AI MODEL GPT LINK .md
Hey everyone,

I've developed a new AI model for geoinformatics students. This model is built on the Chat GPT framework and is designed to assist with various geoinformatics queries and task codes. It's designed for machine learning, automation, GIS tools like QGIS, ArcGIS, and ERDAS, as well as other programming languages and libraries. It's currently in the learning phase, so it might not always provide the correct answers immediately, so provide him with feedback, and he will learn and train himself. 

Provide your feedback and errors to help the model learn and train itself.

Please use it, test it, train it with your questions, and share your suggestions: [Geoinformatics Assistant](https://chatgpt.com/g/g-wKWlZcWes-geoinformatics-assistant).

Thanks for your support!

Best regards,  
Shishir Singh